\name{cggRcalibration-package}
\alias{cggRcalibration-package}
\alias{cggRcalibration}
\docType{package}
\title{
cggRcalibration: package for Cerioli MCD outlier detection project
}
\description{
Implements the outlier detection methodology of Cerioli (2010).
}
\details{
\tabular{ll}{
Package: \tab cggRcalibration\cr
Type: \tab Package\cr
Version: \tab 1.0.12\cr
}
}
\author{
Written and maintained by Christopher G. Green <christopher.g.green@gmail.com>
}
\references{
Andrea Cerioli. Multivariate outlier detection with high-breakdown estimators. 
Journal of the American Statistical Association, 105(489):147–156, 2010.

Andrea Cerioli, Marco Riani, and Anthony C. Atkinson. Controlling the size of multivariate
outlier tests with the mcd estimator of scatter. Statistical Computing, 19:341–353, 2009.
}
\keyword{ package }
\seealso{
}
