\name{CerioliOutlierDetection-package}
\alias{CerioliOutlierDetection-package}
\alias{CerioliOutlierDetection}
\docType{package}
\title{
CerioliOutlierDetection: package for implementing Cerioli (2010) outlier detection method
}
\description{
Implements the outlier detection methodology of Cerioli (2010) based on 
Mahalanobis distances and the minimum covariance determinant (MCD) estimate 
of dispersion.
}
\details{
  \tabular{ll}{
    Package: \tab CerioliOutlierDetection\cr
    Type: \tab Package\cr
    Version: \tab 1.0.0\cr
  }
}
\author{
Written and maintained by Christopher G. Green <christopher.g.green@gmail.com>
}
\references{
Andrea Cerioli. Multivariate outlier detection with high-breakdown estimators. 
Journal of the American Statistical Association, 105(489):147-156, 2010.
}
\keyword{ package }
